+++ GHEtool/VariableClasses/LoadData/_LoadData.py
@@ -12,7 +12,7 @@ class _LoadData(BaseClass, ABC):
     This class contains information w.r.t. load data for the borefield sizing.
     """
     __slots__ = 'hourly_resolution', 'simulation_period', '_peak_heating_duration', '_peak_cooling_duration', 'tm',\
-                '_all_months_equal'
+                '_all_months_equal', '_dhw_yearly'
 
     AVG_UPM: float = 730.  # number of hours per month
     DEFAULT_LENGTH_PEAK: int = 6  # hours
@@ -34,6 +34,7 @@ class _LoadData(BaseClass, ABC):
         self._peak_heating_duration: int = _LoadData.DEFAULT_LENGTH_PEAK
         self.tm: int = _LoadData.AVG_UPM * 3600  # time in a month in seconds
         self._all_months_equal: bool = True  # true if it is assumed that all months are of the same length
+        self._dhw_yearly: float = 0.
 
     @property
     def all_months_equal(self) -> bool:
@@ -68,7 +69,7 @@ class _LoadData(BaseClass, ABC):
     @property
     def UPM(self) -> np.ndarray:
         """
-        Depending on whether or not all months are assumed to have equal length, the UPM are either constant
+        Depending on whether all months are assumed to have equal length, the UPM are either constant
         or vary during the year.
 
         Returns
@@ -540,3 +541,52 @@ class _LoadData(BaseClass, ABC):
         max peak heating : float
         """
         return np.max(self.peak_heating)
+
+    def add_dhw(self, dhw: float) -> None:
+        """
+        This function adds the domestic hot water (dhw).
+        An error is raies is the dhw is not positive.
+
+        Parameters
+        ----------
+        dhw : float
+            Yearly consumption of domestic hot water [kWh/year]
+
+        Returns
+        -------
+        None
+        """
+        self.dhw = dhw
+
+    @property
+    def dhw(self) -> float:
+        """
+        This function returns the yearly domestic hot water consumption.
+
+        Returns
+        -------
+        dhw : float
+            Yearly domestic hot water consumption [kWh/year]
+        """
+        return self._dhw_yearly
+
+    @dhw.setter
+    def dhw(self, dhw: float) -> None:
+        """
+        This function adds the domestic hot water (dhw).
+        An error is raies is the dhw is not positive.
+
+        Parameters
+        ----------
+        dhw : float
+            Yearly consumption of domestic hot water [kWh/year]
+
+        Returns
+        -------
+        None
+        """
+        if not isinstance(dhw, (float, int)):
+            raise ValueError('Please fill in a number for the domestic hot water.')
+        if not dhw >= 0:
+            raise ValueError(f'Please fill in a positive value for the domestic hot water instead of {dhw}.')
+        self._dhw_yearly = dhw